src/requirements.txt
vk_api
DateTime
python-dotenv